Project: amazon-clone-two. Create the home page and the  product page.
Create the Product component and the Product styles.

Create the Home component which is the home page also. It uses the Product component. It goes with the Home stylesheet file created along with it.

Make a subtle change to the header.style.css file to ajust the search input bar with the search button

In the index.css file we add two stylesheet rule.

Inserted the Home component in the main App component

Add two images file that represent the stars rating